requirements.txt
discord.py==2.5.2
python-dotenv==1.1.0
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.0
langchain-community==0.3.31
faiss-cpu==1.12.0
pydantic==2.12.0
django
fastapi
uvicorn[standard]
django-cors-headers
gunicorn
requests
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
